// apps/api/src/opensearch/dto/search-auctions.dto.d.ts
export declare class SearchAuctionsDto {
    page?: number;
    limit?: number;
    sortBy?: string;
    sortOrder?: 'asc' | 'desc';
    search?: string;
    source?: string[];
    sourceIds?: string;
    ids?: string;
    vin?: string;
    year?: number;
    yearMin?: number;
    yearMax?: number;
    make?: string[];
    model?: string[];
    bodyType?: string[];
    trim?: string[];
    yardName?: string[];
    sellerName?: string[];
    excludeUnknownSellers?: boolean;
    transmission?: string;
    fuelType?: string;
    drivetrain?: string[];
    color?: string[];
    cylinders?: string[];
    sellerCategory?: string[];
    engineSizeMin?: number;
    engineSizeMax?: number;
    zip?: string;
    radiusMiles?: number;
    locationState?: string[];
    odometerMin?: number;
    odometerMax?: number;
    odometerBrand?: string;
    damageDescription?: string[];
    saleStatus?: string;
    saleTitleType?: string[];
    titleCategory?: string[];
    hasKeys?: string;
    runsDrives?: string;
    lotCondCode?: string;
    wholesale?: string;
    saleLight?: string[];
    priceMin?: number;
    priceMax?: number;
    saleDateFrom?: number;
    saleDateTo?: number;
    carfaxCleanTitle?: boolean;
    carfax1Owner?: boolean;
    domMax?: number;
    hasCarfaxReport?: boolean;
    hasBuyItNow?: boolean;
    discarded?: boolean;
    inspectableOnly?: boolean;
    includeAggregations?: boolean;
}